refactor(geode-core): formalize mesh as canonical module home [9/10]
Child [9/10] of epic #377 (geode-core hierarchical namespace). `mesh/`
is already directory-backed, so this child moves no files; it only
deprecates the last non-deprecated flat-root mesh re-exports and
migrates same-crate callers onto `geode_core::mesh::*`.

- lib.rs: add `#[deprecated(note = "use geode_core::mesh::<item> instead")]`
  to the `pub use mesh::{...}` block (37 items) and a flat-root note to
  the `PHYS_VACUUM_BUFFER` shim (keeping `#[allow(deprecated)]` since the
  const is already `#[deprecated]` at source).
- Migrate 12 same-crate `crate::TetMesh`/`crate::cube_tet_mesh` code uses
  (incl. the two in-src `#[cfg(test)]` modules) and 3 intra-doc links to
  `crate::mesh::*`, keeping the lib/tests warning-free under the clippy gate.
- mesh/mod.rs: frame `geode_core::mesh` as the canonical mesh home in
  module rustdoc.

Cross-crate flat-root item uses in tests/benches/examples are left on the
shims and deferred to cleanup child #387. `traits` (Mesh/Element/Operator)
and `silvermuller_self_consistent` are untouched by design. Pure refactor,
no behavior change, no file moves.

Closes #386
