+++ crates/geode-core/src/eigen/dense.rs
@@ -299,7 +299,7 @@ pub fn apply_dirichlet_bc(
 /// nodes lying on any of the six faces.
 ///
 /// Tolerance is set tight enough to catch the cube generated by
-/// [`crate::cube_tet_mesh`] (which places nodes at exact `k/n * side`
+/// [`crate::mesh::cube_tet_mesh`] (which places nodes at exact `k/n * side`
 /// coordinates) but loose enough to absorb f64 round-off.
 pub fn cube_interior_mask(nodes: &[[f64; 3]], side: f64) -> Vec<bool> {
     let tol = 1e-9 * side.max(1.0);
